beginning textbox refacto and debugging
