+++ app/javascript/channels/play2interactions.js
@@ -1,0 +1,12 @@
+import { status } from "../scenes/tutorial";
+import { textbox } from '../components/textBox';
+
+var key;
+var next;
+
+export {
+  beginningInstructions,
+  minigameShelves,
+  minigameDoor,
+  endingInstructions,
+};
